+++ src/services/TheaterService.js
@@ -1,29 +1,43 @@
 import $ from 'jquery'
 import axios from 'axios'
-import URLS from '../services/URLS'
+import URLS from './URLS'
+import AccountService from './AccountService'
 
 const urlService = new URLS();
+const accountService = new AccountService();
 
 export default class TheaterService {
     createTheater(theater) {
-        let url = urlService.TheaterURL;
-        axios.post(url, theater)
-        .then((response) => {
-            console.log(response);
-        }).catch(error => {console.log(error)});
+        if (localStorage.token != undefined)
+        {
+            accountService.createHeaders();
+            let url = urlService.TheaterURL;
+            axios.post(url, theater)
+            .then((response) => {
+                console.log(response);
+            }).catch(error => {console.log(error)});
+        }
     }
     editTheater(theater) {
-        let url = urlService.TheaterURL;
-        axios.put(url, theater)
-        .then((response) => {
-            console.log(response);
-        }).catch(error => {console.log(error)});
+        if (localStorage.token != undefined)
+        {
+            accountService.createHeaders();
+            let url = urlService.TheaterURL;
+            axios.put(url, theater)
+            .then((response) => {
+                console.log(response);
+            }).catch(error => {console.log(error)});
+        }
     }
     deleteTheater(id) {
-        let url = urlService.TheaterURL + '/' + id;
-        axios.delete(url)
-        .then((response) => {
-            console.log(response);
-        }).catch(error => {console.log(error)});
+        if (localStorage.token != undefined)
+        {
+            accountService.createHeaders();
+            let url = urlService.TheaterURL + '/' + id;
+            axios.delete(url)
+            .then((response) => {
+                console.log(response);
+            }).catch(error => {console.log(error)});
+        }
     }
 }
